+++ engine/lite/components/request-response/src/main/scala/pl/touk/nussknacker/engine/lite/components/requestresponse/jsonschema/sources/JsonSchemaRequestResponseSourceFactory.scala
@@ -1,14 +1,14 @@
 package pl.touk.nussknacker.engine.lite.components.requestresponse.jsonschema.sources
 
 import org.everit.json.schema.Schema
-import pl.touk.nussknacker.engine.api.{MetaData, NodeId}
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.context.transformation.{NodeDependencyValue, SingleInputGenericNodeTransformation}
 import pl.touk.nussknacker.engine.api.definition.{NodeDependency, TypedNodeDependency}
 import pl.touk.nussknacker.engine.api.process.{BasicContextInitializer, Source}
 import pl.touk.nussknacker.engine.api.typed._
 import pl.touk.nussknacker.engine.api.typed.typing.Unknown
-import pl.touk.nussknacker.engine.json.{JsonSchemaExtractor, JsonSchemaTypeDefinitionExtractor}
+import pl.touk.nussknacker.engine.api.{MetaData, NodeId}
+import pl.touk.nussknacker.engine.json.{JsonSchemaExtractor, SwaggerBasedJsonSchemaTypeDefinitionExtractor}
 import pl.touk.nussknacker.engine.requestresponse.api.openapi.RequestResponseOpenApiSettings.InputSchemaProperty
 import pl.touk.nussknacker.engine.requestresponse.api.{RequestResponseSource, RequestResponseSourceFactory}
 
@@ -28,7 +28,7 @@ class JsonSchemaRequestResponseSourceFactory extends RequestResponseSourceFactor
       val validationResult = determinedSchema.swap.toList.flatMap(_.toList)
       val finalState = determinedSchema.toOption
       val finalInitializer = determinedSchema.toOption.fold(new BasicContextInitializer(Unknown)) { schema =>
-        val schemaTypingResult = JsonSchemaTypeDefinitionExtractor.typeDefinition(schema)
+        val schemaTypingResult = SwaggerBasedJsonSchemaTypeDefinitionExtractor.swaggerType(schema).typingResult
         new BasicContextInitializer(schemaTypingResult)
       }
       FinalResults.forValidation(context, validationResult, finalState)(finalInitializer.validationContext)