+++ engine/lite/kafka/runtime/src/main/scala/pl/touk/nussknacker/engine/lite/requestresponse/RequestResponseConfig.scala
@@ -4,4 +4,4 @@ import pl.touk.nussknacker.engine.requestresponse.openapi.OApiServer
 
 case class OpenApiDefinitionConfig(server: Option[OApiServer] = None)
 
-case class RequestResponseConfig(port: Int, interface: String = "0.0.0.0", definitionMetadata: OpenApiDefinitionConfig = OpenApiDefinitionConfig())
+case class RequestResponseConfig(definitionMetadata: OpenApiDefinitionConfig = OpenApiDefinitionConfig())