+++ utils/default-helpers/src/main/scala/pl/touk/nussknacker/engine/util/functions/numeric.scala
@@ -1,11 +1,48 @@
 package pl.touk.nussknacker.engine.util.functions
 
+import cats.data.{NonEmptyList, ValidatedNel}
+import cats.implicits.catsSyntaxValidatedId
+import pl.touk.nussknacker.engine.api.generics.{GenericFunctionTypingError, GenericType, MethodTypeInfo, Parameter, TypingFunction}
+import pl.touk.nussknacker.engine.api.typed.supertype.NumberTypesPromotionStrategy
+import pl.touk.nussknacker.engine.api.typed.typing
+import pl.touk.nussknacker.engine.api.typed.typing.Typed
 import pl.touk.nussknacker.engine.api.{Documentation, ParamName}
 import pl.touk.nussknacker.engine.util.MathUtils
+import pl.touk.nussknacker.engine.util.functions.numeric.{SingleArgumentMathTypingFunction, LargeNumberOperatorTypingFunction, MinMaxTypingFunction, MathOperatorTypingFunction, ToNumberTypingFunction}
 
 trait numeric extends MathUtils {
+  @GenericType(typingFunction = classOf[MinMaxTypingFunction])
+  override def min(n1: Number, n2: Number): Number = super.min(n1, n2)
+
+  @GenericType(typingFunction = classOf[MinMaxTypingFunction])
+  override def max(n1: Number, n2: Number): Number = super.max(n1, n2)
+
+  @GenericType(typingFunction = classOf[MathOperatorTypingFunction])
+  override def sum(n1: Number, n2: Number): Number = super.sum(n1, n2)
+
+  @GenericType(typingFunction = classOf[LargeNumberOperatorTypingFunction])
+  override def largeSum(n1: Number, n2: Number): Number = super.largeSum(n1, n2)
+
+  @GenericType(typingFunction = classOf[MathOperatorTypingFunction])
+  override def plus(n1: Number, n2: Number): Number = super.plus(n1, n2)
+
+  @GenericType(typingFunction = classOf[MathOperatorTypingFunction])
+  override def minus(n1: Number, n2: Number): Number = super.minus(n1, n2)
+
+  @GenericType(typingFunction = classOf[MathOperatorTypingFunction])
+  override def multiply(n1: Number, n2: Number): Number = super.multiply(n1, n2)
+
+  @GenericType(typingFunction = classOf[MathOperatorTypingFunction])
+  override def divide(n1: Number, n2: Number): Number = super.divide(n1, n2)
+
+  @GenericType(typingFunction = classOf[MathOperatorTypingFunction])
+  override def remainder(n1: Number, n2: Number): Number = super.remainder(n1, n2)
+
+  @GenericType(typingFunction = classOf[SingleArgumentMathTypingFunction])
+  override def negate(n1: Number): Number = super.negate(n1)
 
   @Documentation(description = "Parse string to number")
+  @GenericType(typingFunction = classOf[ToNumberTypingFunction])
   def toNumber(@ParamName("stringOrNumber") stringOrNumber: Any): java.lang.Number = stringOrNumber match {
     case s: CharSequence => new java.math.BigDecimal(s.toString)
     case n: java.lang.Number => n
@@ -15,6 +52,7 @@ trait numeric extends MathUtils {
   def pow(@ParamName("a") a: Double, @ParamName("b") b: Double): Double = Math.pow(a, b)
 
   @Documentation(description = "Returns the absolute value of a value.")
+  @GenericType(typingFunction = classOf[SingleArgumentMathTypingFunction])
   def abs(@ParamName("a") a: Number): Number = a match {
     case n: java.lang.Byte => Math.abs(n.intValue())
     case n: java.lang.Short => Math.abs(n.intValue())
@@ -37,4 +75,42 @@ trait numeric extends MathUtils {
 
 }
 
-object numeric extends numeric
+object numeric extends numeric {
+  class ToNumberTypingFunction extends TypingFunction {
+    override def signatures: Option[NonEmptyList[MethodTypeInfo]] = Some(NonEmptyList.of(
+      MethodTypeInfo.withoutVarargs(Parameter("stringOrNumber", Typed(Typed[String], Typed[Number])) :: Nil, Typed[Number])
+    ))
+
+    override def computeResultType(arguments: List[typing.TypingResult]): ValidatedNel[GenericFunctionTypingError, typing.TypingResult] = {
+      if (arguments.head.canBeSubclassOf(Typed[Number])) arguments.head.withoutValue.validNel
+      else Typed[Number].validNel
+    }
+  }
+
+  private class SingleArgumentMathTypingFunction extends TypingFunction {
+    override def computeResultType(arguments: List[typing.TypingResult]): ValidatedNel[GenericFunctionTypingError, typing.TypingResult] =
+      arguments.head.withoutValue.validNel
+  }
+
+  private trait OperatorTypingFunction extends TypingFunction {
+    protected def numberTypesPromotionStrategy: NumberTypesPromotionStrategy
+
+    override def computeResultType(arguments: List[typing.TypingResult]): ValidatedNel[GenericFunctionTypingError, typing.TypingResult] =
+      numberTypesPromotionStrategy.promote(arguments.head, arguments(1)).withoutValue.validNel
+  }
+
+  private class MinMaxTypingFunction extends OperatorTypingFunction {
+    override protected def numberTypesPromotionStrategy: NumberTypesPromotionStrategy =
+      NumberTypesPromotionStrategy.ForMinMax
+  }
+
+  private class MathOperatorTypingFunction extends OperatorTypingFunction {
+    override protected def numberTypesPromotionStrategy: NumberTypesPromotionStrategy =
+      NumberTypesPromotionStrategy.ForMathOperation
+  }
+
+  private class LargeNumberOperatorTypingFunction extends OperatorTypingFunction {
+    override protected def numberTypesPromotionStrategy: NumberTypesPromotionStrategy =
+      NumberTypesPromotionStrategy.ForLargeNumbersOperation
+  }
+}