+++ utils/json-utils/src/main/scala/pl/touk/nussknacker/engine/json/swagger/SwaggerTyped.scala
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.models.media.{ArraySchema, MapSchema, ObjectSchema, Sch
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedNull, TypedObjectTypingResult, TypingResult}
 import pl.touk.nussknacker.engine.json.swagger.parser.{PropertyName, SwaggerRefSchemas}
 
-import java.time.LocalDateTime
+import java.time.{LocalDate, LocalDateTime, LocalTime}
 import scala.annotation.tailrec
 import scala.collection.JavaConverters._
 
@@ -29,6 +29,10 @@ case object SwaggerBigDecimal extends SwaggerTyped
 
 case object SwaggerDateTime extends SwaggerTyped
 
+case object SwaggerDate extends SwaggerTyped
+
+case object SwaggerTime extends SwaggerTyped
+
 case class SwaggerEnum(values: List[String]) extends SwaggerTyped
 
 case class SwaggerArray(elementType: SwaggerTyped) extends SwaggerTyped
@@ -45,19 +49,22 @@ object SwaggerTyped {
     case _ => Option(schema.get$ref()) match {
       case Some(ref) =>
         SwaggerTyped(swaggerRefSchemas(ref), swaggerRefSchemas)
-      case None => (extractType(schema).get, Option(schema.getFormat)) match {
-        case ("object", _) => SwaggerObject(schema.asInstanceOf[ObjectSchema], swaggerRefSchemas)
-        case ("boolean", _) => SwaggerBool
-        case ("string", Some("date-time")) => SwaggerDateTime
-        case ("string", _) => Option(schema.getEnum) match {
+      case None => (extractType(schema), Option(schema.getFormat)) match {
+        case (None, _) => SwaggerObject(schema.asInstanceOf[Schema[Object@unchecked]], swaggerRefSchemas)
+        case (Some("object"), _) => SwaggerObject(schema.asInstanceOf[ObjectSchema], swaggerRefSchemas)
+        case (Some("boolean"), _) => SwaggerBool
+        case (Some("string"), Some("date-time")) => SwaggerDateTime
+        case (Some("string"), Some("date")) => SwaggerDate
+        case (Some("string"), Some("time")) => SwaggerTime
+        case (Some("string"), _) => Option(schema.getEnum) match {
           case Some(values) => SwaggerEnum(values.asScala.map(_.toString).toList)
           case None => SwaggerString
         }
-        case ("integer", _) => SwaggerLong
-        case ("number", None) => SwaggerBigDecimal
-        case ("number", Some("double")) => SwaggerDouble
-        case ("number", Some("float")) => SwaggerDouble
-        case (null, None) => SwaggerNull
+        case (Some("integer"), _) => SwaggerLong
+        case (Some("number"), None) => SwaggerBigDecimal
+        case (Some("number"), Some("double")) => SwaggerDouble
+        case (Some("number"), Some("float")) => SwaggerDouble
+        case (Some("null"), None) => SwaggerNull
         //todo handle unions
         case (typeName, format) => throw new Exception(s"Type $typeName in format: $format, is not supported")
       }
@@ -88,6 +95,10 @@ object SwaggerTyped {
       Typed.typedClass[java.math.BigDecimal]
     case SwaggerDateTime =>
       Typed.typedClass[LocalDateTime]
+    case SwaggerDate =>
+      Typed.typedClass[LocalDate]
+    case SwaggerTime =>
+      Typed.typedClass[LocalTime]
     case SwaggerNull =>
       TypedNull
   }