+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/encode/JsonSchemaOutputValidatorPrinter.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.schemedkafka.encode
 
 import org.everit.json.schema.{ObjectSchema, Schema, StringSchema}
-import pl.touk.nussknacker.engine.json.JsonSchemaTypeDefinitionExtractor
+import pl.touk.nussknacker.engine.json.SwaggerBasedJsonSchemaTypeDefinitionExtractor
 import pl.touk.nussknacker.engine.util.output.OutputValidatorErrorsMessageFormatter
 
 import java.time.{LocalDate, LocalDateTime, LocalTime}
@@ -10,6 +10,7 @@ import scala.collection.JavaConverters
 object JsonSchemaOutputValidatorPrinter {
 
   import OutputValidatorErrorsMessageFormatter._
+
   import JavaConverters._
 
   private implicit class ListTypesPrinter(list: List[String]) {
@@ -29,8 +30,8 @@ object JsonSchemaOutputValidatorPrinter {
     (defaultDisplayType ::: logicalTypeDisplayType.toList).printType
   }
 
-  private def baseDisplayType(schema: Schema) = JsonSchemaTypeDefinitionExtractor
-    .typeDefinition(schema).display
+  private def baseDisplayType(schema: Schema) = SwaggerBasedJsonSchemaTypeDefinitionExtractor
+    .swaggerType(schema).typingResult.display
 
   //todo: remove duplication - JsonSchemaTypeDefinitionExtractor
   private def printLogicalType(schema: Schema): Option[String] = Option(schema match {