+++ docs/Changelog.md
@@ -7,9 +7,11 @@
 * [#3370](https://github.com/TouK/nussknacker/pull/3370) Feature: scenario node category verification on validation
 * [#3390](https://github.com/TouK/nussknacker/pull/3390) Request-response mode available for k8s deployment
 * [#3392](https://github.com/TouK/nussknacker/pull/3392) Validate scenario before deploy
+* [#3436](https://github.com/TouK/nussknacker/pull/3436) Added types with value to results of operators
 * [#3406](https://github.com/TouK/nussknacker/pull/3406) Scalatest 3.0.8 -> 3.2.10, Scalacheck 1.14.0 -> 1.15.0
 * [#3401](https://github.com/TouK/nussknacker/pull/3401) Request-response mode publishes OpenApi specification for its services
 * [#3427](https://github.com/TouK/nussknacker/pull/3427) Added components/common/extra,components/lite/extra,.. directories for purpose of easier components adding
+* [#3437](https://github.com/TouK/nussknacker/pull/3437) Switch RR typing to SwaggerBasedJsonSchemaTypeDefinitionExtractor
 
 1.5.0 (16 Aug 2022)
 ------------------------